Fix price amount precision bug

+++ server/src/main/java/sageassistant/model/FinancialBalance.java
@@ -1,6 +1,7 @@
 package sageassistant.model;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.serializer.PascalNameFilter;
@@ -14,45 +15,45 @@ public class FinancialBalance implements Serializable {
 	private String AccountNO;
 	private String Currency;
 	private String Year;
-	private Float B0;
-	private Float B1;
-	private Float B2;
-	private Float B3;
-	private Float B4;
-	private Float B5;
-	private Float B6;
-	private Float B7;
-	private Float B8;
-	private Float B9;
-	private Float B10;
-	private Float B11;
-	private Float B12;
-	private Float C0;
-	private Float C1;
-	private Float C2;
-	private Float C3;
-	private Float C4;
-	private Float C5;
-	private Float C6;
-	private Float C7;
-	private Float C8;
-	private Float C9;
-	private Float C10;
-	private Float C11;
-	private Float C12;
-	private Float D0;
-	private Float D1;
-	private Float D2;
-	private Float D3;
-	private Float D4;
-	private Float D5;
-	private Float D6;
-	private Float D7;
-	private Float D8;
-	private Float D9;
-	private Float D10;
-	private Float D11;
-	private Float D12;
+	private BigDecimal B0;
+	private BigDecimal B1;
+	private BigDecimal B2;
+	private BigDecimal B3;
+	private BigDecimal B4;
+	private BigDecimal B5;
+	private BigDecimal B6;
+	private BigDecimal B7;
+	private BigDecimal B8;
+	private BigDecimal B9;
+	private BigDecimal B10;
+	private BigDecimal B11;
+	private BigDecimal B12;
+	private BigDecimal C0;
+	private BigDecimal C1;
+	private BigDecimal C2;
+	private BigDecimal C3;
+	private BigDecimal C4;
+	private BigDecimal C5;
+	private BigDecimal C6;
+	private BigDecimal C7;
+	private BigDecimal C8;
+	private BigDecimal C9;
+	private BigDecimal C10;
+	private BigDecimal C11;
+	private BigDecimal C12;
+	private BigDecimal D0;
+	private BigDecimal D1;
+	private BigDecimal D2;
+	private BigDecimal D3;
+	private BigDecimal D4;
+	private BigDecimal D5;
+	private BigDecimal D6;
+	private BigDecimal D7;
+	private BigDecimal D8;
+	private BigDecimal D9;
+	private BigDecimal D10;
+	private BigDecimal D11;
+	private BigDecimal D12;
 	
 	public String getAccountNO() {
 		return AccountNO;
@@ -72,238 +73,238 @@ public class FinancialBalance implements Serializable {
 	public void setYear(String year) {
 		Year = year;
 	}
-	public Float getB0() {
+	public BigDecimal getB0() {
 		return B0;
 	}
-	public void setB0(Float b0) {
+	public void setB0(BigDecimal b0) {
 		B0 = b0;
 	}
-	public Float getB1() {
+	public BigDecimal getB1() {
 		return B1;
 	}
-	public void setB1(Float b1) {
+	public void setB1(BigDecimal b1) {
 		B1 = b1;
 	}
-	public Float getB2() {
+	public BigDecimal getB2() {
 		return B2;
 	}
-	public void setB2(Float b2) {
+	public void setB2(BigDecimal b2) {
 		B2 = b2;
 	}
-	public Float getB3() {
+	public BigDecimal getB3() {
 		return B3;
 	}
-	public void setB3(Float b3) {
+	public void setB3(BigDecimal b3) {
 		B3 = b3;
 	}
-	public Float getB4() {
+	public BigDecimal getB4() {
 		return B4;
 	}
-	public void setB4(Float b4) {
+	public void setB4(BigDecimal b4) {
 		B4 = b4;
 	}
-	public Float getB5() {
+	public BigDecimal getB5() {
 		return B5;
 	}
-	public void setB5(Float b5) {
+	public void setB5(BigDecimal b5) {
 		B5 = b5;
 	}
-	public Float getB6() {
+	public BigDecimal getB6() {
 		return B6;
 	}
-	public void setB6(Float b6) {
+	public void setB6(BigDecimal b6) {
 		B6 = b6;
 	}
-	public Float getB7() {
+	public BigDecimal getB7() {
 		return B7;
 	}
-	public void setB7(Float b7) {
+	public void setB7(BigDecimal b7) {
 		B7 = b7;
 	}
-	public Float getB8() {
+	public BigDecimal getB8() {
 		return B8;
 	}
-	public void setB8(Float b8) {
+	public void setB8(BigDecimal b8) {
 		B8 = b8;
 	}
-	public Float getB9() {
+	public BigDecimal getB9() {
 		return B9;
 	}
-	public void setB9(Float b9) {
+	public void setB9(BigDecimal b9) {
 		B9 = b9;
 	}
-	public Float getB10() {
+	public BigDecimal getB10() {
 		return B10;
 	}
-	public void setB10(Float b10) {
+	public void setB10(BigDecimal b10) {
 		B10 = b10;
 	}
-	public Float getB11() {
+	public BigDecimal getB11() {
 		return B11;
 	}
-	public void setB11(Float b11) {
+	public void setB11(BigDecimal b11) {
 		B11 = b11;
 	}
-	public Float getB12() {
+	public BigDecimal getB12() {
 		return B12;
 	}
-	public void setB12(Float b12) {
+	public void setB12(BigDecimal b12) {
 		B12 = b12;
 	}
-	public Float getC0() {
+	public BigDecimal getC0() {
 		return C0;
 	}
-	public void setC0(Float c0) {
+	public void setC0(BigDecimal c0) {
 		C0 = c0;
 	}
-	public Float getC1() {
+	public BigDecimal getC1() {
 		return C1;
 	}
-	public void setC1(Float c1) {
+	public void setC1(BigDecimal c1) {
 		C1 = c1;
 	}
-	public Float getC2() {
+	public BigDecimal getC2() {
 		return C2;
 	}
-	public void setC2(Float c2) {
+	public void setC2(BigDecimal c2) {
 		C2 = c2;
 	}
-	public Float getC3() {
+	public BigDecimal getC3() {
 		return C3;
 	}
-	public void setC3(Float c3) {
+	public void setC3(BigDecimal c3) {
 		C3 = c3;
 	}
-	public Float getC4() {
+	public BigDecimal getC4() {
 		return C4;
 	}
-	public void setC4(Float c4) {
+	public void setC4(BigDecimal c4) {
 		C4 = c4;
 	}
-	public Float getC5() {
+	public BigDecimal getC5() {
 		return C5;
 	}
-	public void setC5(Float c5) {
+	public void setC5(BigDecimal c5) {
 		C5 = c5;
 	}
-	public Float getC6() {
+	public BigDecimal getC6() {
 		return C6;
 	}
-	public void setC6(Float c6) {
+	public void setC6(BigDecimal c6) {
 		C6 = c6;
 	}
-	public Float getC7() {
+	public BigDecimal getC7() {
 		return C7;
 	}
-	public void setC7(Float c7) {
+	public void setC7(BigDecimal c7) {
 		C7 = c7;
 	}
-	public Float getC8() {
+	public BigDecimal getC8() {
 		return C8;
 	}
-	public void setC8(Float c8) {
+	public void setC8(BigDecimal c8) {
 		C8 = c8;
 	}
-	public Float getC9() {
+	public BigDecimal getC9() {
 		return C9;
 	}
-	public void setC9(Float c9) {
+	public void setC9(BigDecimal c9) {
 		C9 = c9;
 	}
-	public Float getC10() {
+	public BigDecimal getC10() {
 		return C10;
 	}
-	public void setC10(Float c10) {
+	public void setC10(BigDecimal c10) {
 		C10 = c10;
 	}
-	public Float getC11() {
+	public BigDecimal getC11() {
 		return C11;
 	}
-	public void setC11(Float c11) {
+	public void setC11(BigDecimal c11) {
 		C11 = c11;
 	}
-	public Float getC12() {
+	public BigDecimal getC12() {
 		return C12;
 	}
-	public void setC12(Float c12) {
+	public void setC12(BigDecimal c12) {
 		C12 = c12;
 	}
-	public Float getD0() {
+	public BigDecimal getD0() {
 		return D0;
 	}
-	public void setD0(Float d0) {
+	public void setD0(BigDecimal d0) {
 		D0 = d0;
 	}
-	public Float getD1() {
+	public BigDecimal getD1() {
 		return D1;
 	}
-	public void setD1(Float d1) {
+	public void setD1(BigDecimal d1) {
 		D1 = d1;
 	}
-	public Float getD2() {
+	public BigDecimal getD2() {
 		return D2;
 	}
-	public void setD2(Float d2) {
+	public void setD2(BigDecimal d2) {
 		D2 = d2;
 	}
-	public Float getD3() {
+	public BigDecimal getD3() {
 		return D3;
 	}
-	public void setD3(Float d3) {
+	public void setD3(BigDecimal d3) {
 		D3 = d3;
 	}
-	public Float getD4() {
+	public BigDecimal getD4() {
 		return D4;
 	}
-	public void setD4(Float d4) {
+	public void setD4(BigDecimal d4) {
 		D4 = d4;
 	}
-	public Float getD5() {
+	public BigDecimal getD5() {
 		return D5;
 	}
-	public void setD5(Float d5) {
+	public void setD5(BigDecimal d5) {
 		D5 = d5;
 	}
-	public Float getD6() {
+	public BigDecimal getD6() {
 		return D6;
 	}
-	public void setD6(Float d6) {
+	public void setD6(BigDecimal d6) {
 		D6 = d6;
 	}
-	public Float getD7() {
+	public BigDecimal getD7() {
 		return D7;
 	}
-	public void setD7(Float d7) {
+	public void setD7(BigDecimal d7) {
 		D7 = d7;
 	}
-	public Float getD8() {
+	public BigDecimal getD8() {
 		return D8;
 	}
-	public void setD8(Float d8) {
+	public void setD8(BigDecimal d8) {
 		D8 = d8;
 	}
-	public Float getD9() {
+	public BigDecimal getD9() {
 		return D9;
 	}
-	public void setD9(Float d9) {
+	public void setD9(BigDecimal d9) {
 		D9 = d9;
 	}
-	public Float getD10() {
+	public BigDecimal getD10() {
 		return D10;
 	}
-	public void setD10(Float d10) {
+	public void setD10(BigDecimal d10) {
 		D10 = d10;
 	}
-	public Float getD11() {
+	public BigDecimal getD11() {
 		return D11;
 	}
-	public void setD11(Float d11) {
+	public void setD11(BigDecimal d11) {
 		D11 = d11;
 	}
-	public Float getD12() {
+	public BigDecimal getD12() {
 		return D12;
 	}
-	public void setD12(Float d12) {
+	public void setD12(BigDecimal d12) {
 		D12 = d12;
 	}
