+++ server/src/main/java/sageassistant/model/RptInvoice.java
@@ -1,5 +1,6 @@
 package sageassistant.model;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 
 import com.alibaba.fastjson.JSON;
@@ -12,10 +13,10 @@ public class RptInvoice {
 	private String CUR_0;
 	private String NUM_0;
 	private Date ACCDAT_0;
-	private Double AMTATI_0;
-	private Double AMTNOT_0;
-	private Double AMTTAX_0;
-	private Double RATMLT_3;
+	private BigDecimal AMTATI_0;
+	private BigDecimal AMTNOT_0;
+	private BigDecimal AMTTAX_0;
+	private BigDecimal RATMLT_3;
 	private String BPR_0;
 	private String YFAPIAO_0;
 	private String YSIH_CUSDOC_0;
@@ -46,9 +47,9 @@ public class RptInvoice {
 	private String ECCVALMAJ_0;
 	private String ITMDES_0;
 	private String ITMDES2_0;
-	private Float QTY_0;
+	private Double QTY_0;
 	private String SAU_0;
-	private Float NETPRI_0;
+	private BigDecimal NETPRI_0;
 	private String BPIEECNUM_0;
 	private String EECICT_0;
 	private String ICTCTY_0;
@@ -76,11 +77,11 @@ public class RptInvoice {
 	private String YSOQ_BPCPN_0;
 	private Integer YSOQ_ITEMCUS_0;
 	private String PCK_0;
-	private Float YPCKWID_0;
-	private Float YPCKLEN_0;
-	private Float YPCKHEI_0;
+	private Double YPCKWID_0;
+	private Double YPCKLEN_0;
+	private Double YPCKHEI_0;
 	private String WEU_PCK;
-	private Float NETWEI_SPK;
+	private Double NETWEI_SPK;
 	private String PACNUM_0;
 	private String TEXT_DEB;
 	private String TEXT_DET;
@@ -113,28 +114,28 @@ public class RptInvoice {
 	public void setACCDAT_0(Date aCCDAT_0) {
 		ACCDAT_0 = aCCDAT_0;
 	}
-	public Double getAMTATI_0() {
+	public BigDecimal getAMTATI_0() {
 		return AMTATI_0;
 	}
-	public void setAMTATI_0(Double aMTATI_0) {
+	public void setAMTATI_0(BigDecimal aMTATI_0) {
 		AMTATI_0 = aMTATI_0;
 	}
-	public Double getAMTNOT_0() {
+	public BigDecimal getAMTNOT_0() {
 		return AMTNOT_0;
 	}
-	public void setAMTNOT_0(Double aMTNOT_0) {
+	public void setAMTNOT_0(BigDecimal aMTNOT_0) {
 		AMTNOT_0 = aMTNOT_0;
 	}
-	public Double getAMTTAX_0() {
+	public BigDecimal getAMTTAX_0() {
 		return AMTTAX_0;
 	}
-	public void setAMTTAX_0(Double aMTTAX_0) {
+	public void setAMTTAX_0(BigDecimal aMTTAX_0) {
 		AMTTAX_0 = aMTTAX_0;
 	}
-	public Double getRATMLT_3() {
+	public BigDecimal getRATMLT_3() {
 		return RATMLT_3;
 	}
-	public void setRATMLT_3(Double rATMLT_3) {
+	public void setRATMLT_3(BigDecimal rATMLT_3) {
 		RATMLT_3 = rATMLT_3;
 	}
 	public String getBPR_0() {
@@ -311,10 +312,10 @@ public class RptInvoice {
 	public void setITMDES2_0(String iTMDES2_0) {
 		ITMDES2_0 = iTMDES2_0;
 	}
-	public Float getQTY_0() {
+	public Double getQTY_0() {
 		return QTY_0;
 	}
-	public void setQTY_0(Float qTY_0) {
+	public void setQTY_0(Double qTY_0) {
 		QTY_0 = qTY_0;
 	}
 	public String getSAU_0() {
@@ -323,10 +324,10 @@ public class RptInvoice {
 	public void setSAU_0(String sAU_0) {
 		SAU_0 = sAU_0;
 	}
-	public Float getNETPRI_0() {
+	public BigDecimal getNETPRI_0() {
 		return NETPRI_0;
 	}
-	public void setNETPRI_0(Float nETPRI_0) {
+	public void setNETPRI_0(BigDecimal nETPRI_0) {
 		NETPRI_0 = nETPRI_0;
 	}
 	public String getBPIEECNUM_0() {
@@ -491,22 +492,22 @@ public class RptInvoice {
 	public void setPCK_0(String pCK_0) {
 		PCK_0 = pCK_0;
 	}
-	public Float getYPCKWID_0() {
+	public Double getYPCKWID_0() {
 		return YPCKWID_0;
 	}
-	public void setYPCKWID_0(Float yPCKWID_0) {
+	public void setYPCKWID_0(Double yPCKWID_0) {
 		YPCKWID_0 = yPCKWID_0;
 	}
-	public Float getYPCKLEN_0() {
+	public Double getYPCKLEN_0() {
 		return YPCKLEN_0;
 	}
-	public void setYPCKLEN_0(Float yPCKLEN_0) {
+	public void setYPCKLEN_0(Double yPCKLEN_0) {
 		YPCKLEN_0 = yPCKLEN_0;
 	}
-	public Float getYPCKHEI_0() {
+	public Double getYPCKHEI_0() {
 		return YPCKHEI_0;
 	}
-	public void setYPCKHEI_0(Float yPCKHEI_0) {
+	public void setYPCKHEI_0(Double yPCKHEI_0) {
 		YPCKHEI_0 = yPCKHEI_0;
 	}
 	public String getWEU_PCK() {
@@ -515,10 +516,10 @@ public class RptInvoice {
 	public void setWEU_PCK(String wEU_PCK) {
 		WEU_PCK = wEU_PCK;
 	}
-	public Float getNETWEI_SPK() {
+	public Double getNETWEI_SPK() {
 		return NETWEI_SPK;
 	}
-	public void setNETWEI_SPK(Float nETWEI_SPK) {
+	public void setNETWEI_SPK(Double nETWEI_SPK) {
 		NETWEI_SPK = nETWEI_SPK;
 	}
 	public String getPACNUM_0() {
